lara/state: port Lara_State_Initialise to TRX
This moves Lara_State_Initialise to TRX, and also exposes responsive
animation tests for TR1.
