+++ src/libtrx/include/libtrx/game/lara/state.h
@@ -5,4 +5,6 @@ void Lara_State_Register(
     LARA_STATE state, void (*handle_func)(ITEM *item, COLL_INFO *coll));
 void Lara_State_RegisterExtra(
     LARA_EXTRA_STATE state, void (*handle_func)(ITEM *item, COLL_INFO *coll));
+void Lara_State_Initialise(void);
+bool Lara_State_IsResponsive(LARA_ANIMATION anim_idx);
 void Lara_State_Update(ITEM *item, COLL_INFO *coll);
